+++ Functions/functions/studio/calling/options.private.js
@@ -1,6 +1,6 @@
 module.exports = {
-    sayOptions: { voice: 'Polly.Conchita', language: 'es-ES' },
-    holdMusicUrl: '/assets/guitar_music.mp3',
+    sayOptions: { voice: 'Google.es-US-Neural2-A', language: 'es-ES' },
+    holdMusicUrl: '/guitar_music.mp3',
     // Enable Estimated Waiting Time in voice prompt
     getEwt: true,
     //  Time interval (minutes) for Estimated Waiting Time stats
@@ -10,11 +10,11 @@ module.exports = {
     // Priority for the Task generatared by the VoiceMail
     VoiceMailTaskPriority: 50,
     // Agent audible alert sound file for voice mail
-    VoiceMailAlertTone: '/assets/alertTone.mp3',
+    VoiceMailAlertTone: '/alertTone.mp3',
     // Priority for the Task generatared by the VoiceMail
     CallbackTaskPriority: 50,
     // Agent audible alert sound file for callback call
-    CallbackAlertTone: '/assets/alertTone.mp3',
+    CallbackAlertTone: '/alertTone.mp3',
     // Timezone configuration
-    TimeZone: 'America/Los_Angeles',
+    TimeZone: 'America/Bogota',
 };