+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Actualiza URL Servicio Manager de Flex
 * Corrige paths de funciones serverless
 * Corrige path de funciones en directorios
+* Corrige url de assets
 
 ## 27/11/2023
 * Agrega funciones de Admin y plugin Admin